backend/requirements.txt
# FastAPI and ASGI
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Strands Agents SDK
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# Utilities
pydantic[email]>=2.6.0
pydantic-settings>=2.5.2
email-validator>=2.0.0
python-dotenv==1.0.1
httpx>=0.27.1
aiofiles==23.2.1
structlog==24.1.0

# WebSockets
websockets==12.0

# Redis for caching/sessions
redis==5.0.1

# Celery for async tasks
celery[redis]==5.3.6
